TeleopV10! We made it!
This is the new mappings that Joe wanted from the meeting on Thursday.
It also features simplified code. With this done Ethan should spend the
rest of his time DIRECTLY ON AUTONOMOUS with Joe. HEAR ME!? Not sure
about the deleted text file part of thie commit. I didn't delete
anything, lol.